Fix class refactoring to fix build.
In previous iterations of the library, Location, Party, and Role were there own
independent classes in the gov.nist.secauto.oscal.lib.model package namespace,
but previous refactoring moved it into the namespace of the defining metaschema.

They must now be prefixed with the name of the metaschema for the assemblies in
this particular model, Metadata, so import gov.nist.secauto.oscal.lib.model.Metadata.
