+++ src/main/java/gov/nist/secauto/oscal/lib/profile/resolver/support/BasicIndexer.java
@@ -37,10 +37,10 @@ import gov.nist.secauto.oscal.lib.model.BackMatter.Resource;
 import gov.nist.secauto.oscal.lib.model.CatalogGroup;
 import gov.nist.secauto.oscal.lib.model.Control;
 import gov.nist.secauto.oscal.lib.model.ControlPart;
-import gov.nist.secauto.oscal.lib.model.Location;
+import gov.nist.secauto.oscal.lib.model.Metadata.Location;
 import gov.nist.secauto.oscal.lib.model.Parameter;
-import gov.nist.secauto.oscal.lib.model.Party;
-import gov.nist.secauto.oscal.lib.model.Role;
+import gov.nist.secauto.oscal.lib.model.Metadata.Party;
+import gov.nist.secauto.oscal.lib.model.Metadata.Role;
 import gov.nist.secauto.oscal.lib.profile.resolver.ProfileResolver;
 import gov.nist.secauto.oscal.lib.profile.resolver.support.IEntityItem.ItemType;
